Linted

+++ navground_examples/src/collision/cached_collision.cpp
@@ -22,12 +22,12 @@ int main([[maybe_unused]] int argc, [[maybe_unused]] char* argv[]) {
   f.setup({}, 0.0, {}, {Disc{{2.0, 0.0}, 1.0}}, {});
   auto angles = f.get_angles_for_sector(-1.0, 2.0, 3);
   auto distances = f.get_free_distance_for_sector(-1.0, 2.0, 3, 3.0, false);
-  for (int i = 0; i < angles.size(); ++i) {
+  for (size_t i = 0; i < angles.size(); ++i) {
     printf("%.3f: %.3f\n", angles[i], distances[i]);
   }
   printf("Cached\n");
   distances = f.get_free_distance(false);
-  for (int i = 0; i < angles.size(); ++i) {
+  for (size_t i = 0; i < angles.size(); ++i) {
     printf("%.3f: %.3f\n", angles[i], distances[i]);
   }
   return 0;
